- fixes for importing wavs - make separate audio previewer

+++ SCICompanionLib/Src/Resources/ResourceContainer.cpp
@@ -100,7 +100,7 @@ sci::istream ResourceContainer::ResourceIterator::_GetResourceHeaderAndPackage(R
 
     sci::istream temp = (*_container->_mapAndVolumes)[_state.mapIndex]->GetHeaderAndPositionedStream(_currentEntry, rh);
 
-    assert(rh.Number == _currentEntry.Number && "Corrupt resource map");
+    assert((uint16_t)rh.Number == _currentEntry.Number && "Corrupt resource map");
     assert(rh.PackageHint == _currentEntry.PackageNumber && "Corrupt resource map");
     // By setting these to those in the resource map (instead of the header), we can ensure that the ResourceBlob matches
     // the resource map information. This ensures that we can delete resources in the case of a corrupt resource map/package.
